requirements.txt
click==8.1.3
python-decouple==3.6
requests==2.28.1
Django==4.1.*
django-extensions==3.2.*
psycopg2-binary==2.9.*
gunicorn==20.1.0
whitenoise==6.2.0
autopep8==2.0.0
isort==5.10.1
djhtml==1.5.2
django-widget-tweaks==1.4.*
six==1.16.0
openpyxl==3.0.*
rich==12.6.0
django-seed==0.3.1
Pillow==9.3.0
ipdb
ruff==0.0.254
faker-commerce==1.0.3
django-import-export==3.1.0
pandas==1.5.3
dask==2023.3.1
PyExcelerate==0.10.0
